+++ service/CompanyLogo/UploadPictureService.js
@@ -1,6 +1,6 @@
 const axios = require('axios');
 const config = require('../../config/config');
-const scrapeWebsite = require('../webScrapeService');
+const scrapeWebsite = require('../ScrapeJobs/webScrapeService');
 const writePost = require('../../helpers/writePostContent');
 const fs = require('fs');
 const formData = require('form-data');
